Working on eventPage
