+++ planit-react/src/pages/EventPage/EventPage.module.css
@@ -1,0 +1,13 @@
+/* EventPage.module.css */
+.centered {
+  display: flex;
+  justify-content: center;
+}
+
+.image {
+  display: block;
+  margin: 0 auto;
+  max-width: 50%;
+  height: auto;
+  margin-top: 10px;
+}
